script from time.js
