+++ homeworks/anna.vrana_Anna1o1Vrana/4-modules/index.html
@@ -28,6 +28,6 @@
     <div class="time-wrapper">
     </div>
 </div>
-<script src="scripts/main.js"></script>
+<script type="module" src="scripts/main.js"></script>
 </body>
 </html>
